requirements.txt
Django>=5.0,<6.0
psutil>=5.9
pycryptodome>=3.20
